split main.cpp per platform
