+++ loader/src/load.hpp
@@ -1,0 +1,2 @@
+#pragma once
+int geodeEntry(void* platformData);
